requirements.txt
google-generativeai>=0.3.0
streamlit>=1.24.0
python-dotenv>=0.19.0
markdown>=3.4.3
pygments>=2.15.1
pytest>=7.3.1
black>=23.3.0
pylint>=2.17.4
PyPDF2>=3.0.0
pdfplumber>=0.10.2
logging-config>=1.0.3
sentry-sdk>=1.39.1
python-docx>=0.8.11
gunicorn>=21.2.0
redis>=5.0.1
rq>=1.15.1
reportlab>=4.0.4
textblob>=0.17.1
nltk>=3.8.1
requests>=2.31.0
beautifulsoup4>=4.12.2
radon>=6.0.1
autopep8>=2.0.4
graphviz>=0.20.1
pandas>=2.0.0 